Adding generic row to dataclass functionality #14
