+++ src/main/kotlin/bsu/cc/views/MainView.kt
@@ -1,7 +1,8 @@
 package bsu.cc.views
 
 import bsu.cc.Styles
-import bsu.cc.parser.ParserDemo
+import bsu.cc.parser.DemoParser
+import bsu.cc.parser.XlsxParser
 import javafx.scene.control.Alert.AlertType.INFORMATION
 import javafx.scene.control.TextField
 import tornadofx.*
@@ -50,7 +51,7 @@ class MainView : View("Conflict Checker") {
                 }
                 button("Print Test") {
                     setOnAction{
-                        ParserDemo.demoParse(fileNameField.text)
+                        DemoParser.demoParse(fileNameField.text)
                     }
                 }
             }
